Added transaction functionality to portfolios

+++ src/main/java/com/springtrader/controller/PortfolioController.java
@@ -9,6 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.core.Authentication;
+import org.springframework.ui.Model;
 
 import com.springtrader.model.portfolio.NewPortfolio;
 import com.springtrader.model.portfolio.Portfolio;
@@ -32,14 +34,26 @@ public class PortfolioController {
       return portfolioService.addPortfolio(userName, portfolio.getPortfolioName());
    }
    
+   @RequestMapping("/users/{username}/portfolios/{id}")
+   public List<TradeRow> showTransactions(@PathVariable("username") String username, @PathVariable("id") long id) {
+      return portfolioService.getTransactions(username, id);
+   }
+   
    @DeleteMapping("/users/{username}/portfolios/{id}")
-   public String deletePortfolio(@PathVariable("username") String username, @PathVariable("id") long id) {
-      return portfolioService.deletePortfolio(id, username);
+   public String deletePortfolio(@PathVariable("username") String username, Authentication auth, @PathVariable("id") long id) {
+	  if(auth.getName().equals(username)) {
+		  return portfolioService.deletePortfolio(id, username);
+	  }
+	  	return "You can't do that...";
+     
    }
    
    @PostMapping("/users/{username}/portfolios/{id}")
-   public String addTransaction(@PathVariable("username") String username, @PathVariable("id") long id, @RequestBody TradeRow transaction) {
-      return portfolioService.addTransaction(id, username, transaction);
+   public String addTransaction(@PathVariable("username") String username, Authentication auth, @PathVariable("id") long id, @RequestBody TradeRow transaction) {
+	   if(auth.getName().equals(username)) {
+		      return portfolioService.addTransaction(id, username, transaction);
+		  }
+		  	return "You can't do that...";
    }
